+++ kamehouse-shell/bin/functions/minimal-functions.sh
@@ -9,7 +9,7 @@ SCRIPT_NAME_NO_EXT=${SCRIPT_NAME%.*}
 SCRIPT_START_DATE="$(date +%Y-%m-%d' '%H:%M:%S)"
 SCRIPT_START_TIME="$(date +%s)"
 # Script configuration file. 
-# This config file can be used to override the default variable values defined in setDefaultScriptConfig() 
+# This config file can be used to override the default variable values defined in initScriptConfig() 
 # and also to override default values of script variables that can be modified by script command line parameters.
 SCRIPT_CONFIG_FILENAME=${SCRIPT_NAME_NO_EXT}.cfg
 SCRIPT_CONFIG_PATH=${HOME}/.kamehouse/config/shell